Give up on state when no path exists.
Prior to this commit when it wasn't possible to trace a path to the
desired next crossing instead of doing the right thing and giving up on
this path it would instead just skip crossings until it found one that
it could trace a path to.  this produced incorrect results.