Add an abstract method for determining reality of a tangle
Previously I was attempting to plot a tangle in order to determine
whether or not it was real. While this is an accurate method it involves
a lot of work that is irrelevant to the simple real/imaginary question.

This commit adds an isReal method to Tangle that performs an abstract
check that verifies that the topology of the tangle is real without any
of the overhead of producing a drawing. Consequently it's a lot faster.