additional commentary for findCanonFirst
This commit simply adds additional commentary to the findCanonFirst
method of Knot::Tangle.

This commentary gives a brief overview of what we'll actually do to find
the "Canon First" using the idea already presented but without having to
do a lot of Base-N math.

It then updates the code in findCanonFirst to use the just-described
method that doesn't have to do any Base-N math.

For small Tangles the difference in the amount of work that needs to be
done will be negligible, but for large Tangles (100+ crossings) the
amount of work that needs to be done will be greatly reduced.